+++ public/js/home/smoothAnchor.js
@@ -1,0 +1,22 @@
+document.querySelectorAll('a[href^="#"]').forEach(anchor => {
+    anchor.addEventListener('click', function (e) {
+        e.preventDefault();
+
+        document.querySelector(this.getAttribute('href')).scrollIntoView({
+            behavior: 'smooth'
+        });
+    });
+});
+
+var mybutton = document.getElementById("return_up");
+window.onscroll = function() {scrollFunction()};
+
+function scrollFunction() {
+    if (document.body.scrollTop > 250 || document.documentElement.scrollTop > 250) {
+        mybutton.style.display = "block";
+        mybutton.style.opacity = "1"
+    } else {
+        mybutton.style.opacity = "0"
+        mybutton.style.display = "none";
+    }
+  }
