+++ templates/home/index.html.twig
@@ -1,6 +1,6 @@
 {% extends 'base.html.twig' %}
 
-{% block title %}The Front's OnlyFan{% endblock %}
+{% block title %}Speedrun{% endblock %}
 
 {% block javascript %}
 {% endblock %}
@@ -44,7 +44,7 @@
 </div>
 <div id='firstContainer'>
     <img src="https://picsum.photos/1920/1080" alt="fondEcranSite">
-    <h1><span>Welcome to</span><br>FRONT GOD</h1>
+    <h1><span>Welcome to</span><br>SPEEDRUN</h1>
     <a href="#secondContainer"><i class="fas fa-chevron-down"></i></a>
 </div>
 <div id='secondContainer'>
@@ -132,9 +132,9 @@
     {# Icon FontAwesome #}
     <script src="https://kit.fontawesome.com/f953e20ddb.js" crossorigin="anonymous"></script>
     {# SmoothAnchor Transitions #}
-    <script src="{{ asset('js/smoothAnchor.js') }}"></script>
+    <script src="{{ asset('js/home/smoothAnchor.js') }}"></script>
     {# NavBar Script #}
-    <script src="{{ asset('js/hideNavBar.js') }}"></script>
+    <script src="{{ asset('js/home/hideNavBar.js') }}"></script>
     {# AJAX SEARCH Script #}
-    <script src="{{ asset('js/ajaxGameSearch.js') }}"></script>
+    <script src="{{ asset('js/home/ajaxGameSearch.js') }}"></script>
 {% endblock %}
